가독성을 위한 config 폴더 계층 수정
