+++ config/theme.js
@@ -2,7 +2,7 @@ const lightCodeTheme = require('prism-react-renderer/themes/github');
 const darkCodeTheme = require('prism-react-renderer/themes/dracula');
 
 const colorMode = {
-  defaultMode: 'light',
+  defaultMode: 'dark',
   disableSwitch: false,
   respectPrefersColorScheme: true,
 };
